pyproject.toml
[project]
name = "dsat-lightning"
version = "0.1.0"
description = "DSAT with PyTorch Lightning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.2",
    "h5py>=3.15.1",
    "hydra-core>=1.3.2",
    "lightning>=2.6.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "opencv-python-headless>=4.13.0.92",
    "pandas>=3.0.0",
    "polartransform>=2.0.0",
    "rich>=14.3.3",
    "tables>=3.10.2",
    "torch>=2.10.0",
    "wandb>=0.24.2",
]

[project.scripts]
train = "dsat_lightning.train:main"
predict = "dsat_lightning.predict:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dsat_lightning"]

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
